requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
gunicorn
psycopg2
